Created unit test for:
1.CourseModeltest:  self.assertAlmostEqual(course.created_at, now) (previously created)

2.StepModelTest : self.assertIn(step, self.course.step_set.all())

3.CourseViewsTests : self.assertEqual(resp.status_code, 200)
                     self.assertIn(self.course, resp.context['courses'])
                     self.assertIn(self.course2, resp.context['courses'])
Notes:
django.core.urlresolvers.reverse() takes a URL name and reverses it to the correct URL string. More information

self.client acts like a web browser and lets you make requests to URLs, both inside and outside of your Django project.

assertEqual(a, b) checks that a and b are equal to each other.

assertIn(a, b) checks that a is contained in b.